fix import paths for forms and remove unused route definitions
