+++ app/src/pages/epi/classes.html
@@ -5,13 +5,15 @@
         <div class="card-header">
             <h2 class="m-2">Classes EPI</h1>
                 <hr>
-                <button type="button" class="btn btn-success mb-3" data-bs-toggle="modal" data-bs-target="#cadastroClasses">
+                <button type="button" class="btn btn-success mb-3" data-bs-toggle="modal"
+                    data-bs-target="#CadastroCategorias">
                     <span class="icon text-white-50">
                         <i class="fa-solid fa-plus"></i>
                     </span>
                     <span class="text">Cadastrar</span>
                 </button>
-                <button type="button" class="btn btn-warning mb-3" data-bs-toggle="modal" data-bs-target="#importarquivo">
+                <button type="button" class="btn btn-warning mb-3" data-bs-toggle="modal"
+                    data-bs-target="#importarquivo">
                     <span class="icon text-white-50">
                         <i class="fa-solid fa-file-import"></i>
                     </span>
