+++ app/src/pages/forms/classes/create.html
@@ -1,9 +1,10 @@
-<div class="modal fade" id="cadastroClasses" tabindex="-1" aria-labelledby="cadastroClassesLabel" aria-hidden="true">
+<div class="modal fade" id="CadastroCategorias" tabindex="-1" aria-labelledby="CadastroCategoriasLabel"
+    aria-hidden="true">
     <div class="modal-dialog modal-xl">
         <div class="modal-content">
             <div class="modal-header">
                 <h4 class="m-0 font-weight-bold text-primary">Cadastrar Tipo EPI</h4>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
             </div>
             <div class="modal-body">
                 <form action="{{ url_for('cadastrar', tipo = request.endpoint) }}" method="post">
