+++ app/src/pages/forms/classes/edit.html
@@ -1,10 +1,11 @@
 <script src="{{ url_for('static', filename='js/hideLoad.js') }}"></script>
-<div class="modal fade" data-modal_edit="edit" id="cadastroClasses" tabindex="-1" aria-labelledby="cadastroClassesLabel" aria-hidden="true">
+<div class="modal fade" data-modal_edit="edit" id="CadastroCategorias" tabindex="-1"
+    aria-labelledby="CadastroCategoriasLabel" aria-hidden="true">
     <div class="modal-dialog modal-xl">
         <div class="modal-content">
             <div class="modal-header">
                 <h4 class="m-0 font-weight-bold text-primary">Editar Tipo EPI</h4>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
             </div>
             <div class="modal-body">
                 <form action="{{ url_for('cadastrar', tipo = request.endpoint) }}" method="post">
